Add files via upload
CREATING NEW TRANSFORMERS FOR SCIKIT-LEARN•
You are going to create a new transformer for scikit-learn, similar to the SimpleImputerQuartile() developed in the theory slides.•
1.Create a matrix X with a few instances with some nan values in it.
Create also a response variable vector y. You can put any values in both X and y.•
2.Following the structure explained for SimpleImputerQuartile, define a new transformer that replaces a nan value at location X[i,j] by a random value sampled from a normal distribution with mean mj and standard deviation sj. You can use the numpy.random.normal for sampling random values from a gaussian distribution. Check the numpy.random.normal documentation.•
3.Show that it works for your data matrix (executing the .fit and the .transform methods).•
4.Show that it works in a pipeline with knn (executing the .fit and the .predict methods of the pipeline).•
5.To hand in: 
The code and a report showing that your program works.
Or,a jupyter notebook with comments.